refactor(channel): extract sidebar state into channel_sidebar_state.go
Continues the channel.go decomposition (#514 was the worked example).
Pulls sidebar items projection, cursor management, key handler, and the
sidebarItem type into a dedicated kin file. channel_sidebar.go remains
the rendering side.

  - cmd/wuphf/channel.go: -224 LOC (4687 -> 4463)
  - cmd/wuphf/channel_sidebar_state.go: +245 LOC (new)
  - cmd/wuphf/channel_sidebar_state_test.go: +131 LOC (new)

Methods moved:
  - sidebarItem (type)
  - sidebarItemAt (mouse hit-test)
  - updateSidebar (key handler)
  - sidebarItems / channelSidebarItems / appSidebarItems / quickJumpItems
  - setSidebarCursorForItem / clampSidebarCursor
  - selectSidebarItem (channel + app dispatch)
  - syncSidebarCursorToActive

Tests added (epicenter — one branch each):
  - 1:1 mode hides the sidebar
  - sidebarItems prepends channels before apps
  - empty channel list falls back to #general
  - clampSidebarCursor handles empty / overflow / negative
  - setSidebarCursorForItem finds matching kind+value
  - syncSidebarCursorToActive prefers channel match in messages app

No behavior change. All existing tests pass.
